Organize files in main package into proper subpackages
