+++ app/src/main/java/com/vmware/nimbus/ui/main/viewmodels/DeploymentsViewModel.java
@@ -1,4 +1,4 @@
-package com.vmware.nimbus.ui.main;
+package com.vmware.nimbus.ui.main.viewmodels;
 
 import androidx.lifecycle.ViewModel;
