+++ app/src/main/res/layout/fragment_deployments.xml
@@ -8,7 +8,7 @@
     android:layout_weight="1"
     android:orientation="vertical"
     android:paddingTop="4dp"
-    tools:context=".ui.main.DeploymentsViewFragment">
+    tools:context=".ui.main.fragments.DeploymentsViewFragment">
 
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/fragment_deployments_recycler"
